requirements.txt
newspaper3k==0.2.2
vaderSentiment==2.5
